selftest: side parameter (yes|no); fee truth at sub-cent f64 resolution (demo-observed 0.46c rounded to 0)
Demo live-fill verification: YES 1ct@7c filled (fee 0.0046/ct = formula exact), NO 1ct@73c filled via ask@0.2700 inversion (fee 0.0138/ct exact). All three V2 open risks closed.

+++ crates/engine/src/strategy.rs
@@ -47,7 +47,7 @@ pub struct FillReport {
     /// ACTUAL total fee in cents from the exchange's `average_fee_paid` (None in
     /// paper mode or when nothing filled). Recorded ALONGSIDE our own taker-fee
     /// estimate — real fees are the mechanics-week deliverable.
-    pub actual_fee_cents: Option<i64>,
+    pub actual_fee_cents: Option<f64>,
     /// True for paper-mode simulated fills.
     pub simulated: bool,
     /// Unix-ms timestamps for latency measurement (week-1 deliverable).
